First steps of attribute cache initialization.
This is done to fully understand how attribute initialization works
with the current architecture, while taking care of its particular
difference compared to the handling of exclude patterns.
