+++ gix-worktree/src/fs/cache/state/ignore.rs
@@ -1,23 +1,9 @@
-use std::path::Path;
-
+use crate::fs::cache::state::IgnoreMatchGroup;
+use crate::fs::PathOidMapping;
 use bstr::{BStr, BString, ByteSlice};
 use gix_glob::pattern::Case;
 use gix_hash::oid;
-
-use crate::fs::{cache::State, PathOidMapping};
-
-type AttributeMatchGroup = gix_attributes::Search;
-type IgnoreMatchGroup = gix_ignore::Search;
-
-/// State related to attributes associated with files in the repository.
-#[derive(Default, Clone)]
-#[allow(unused)]
-pub struct Attributes {
-    /// Attribute patterns that match the currently set directory (in the stack).
-    pub stack: AttributeMatchGroup,
-    /// Attribute patterns which aren't tied to the repository root, hence are global. They are consulted last.
-    pub globals: AttributeMatchGroup,
-}
+use std::path::Path;
 
 /// State related to the exclusion of files.
 #[derive(Default, Clone)]
@@ -35,7 +21,7 @@ pub struct Ignore {
     /// (index into match groups, index into list of pattern lists, index into pattern list)
     matched_directory_patterns_stack: Vec<Option<(usize, usize, usize)>>,
     ///  The name of the file to look for in directories.
-    exclude_file_name_for_directories: BString,
+    pub(crate) exclude_file_name_for_directories: BString,
     /// The case to use when matching directories as they are pushed onto the stack. We run them against the exclude engine
     /// to know if an entire path can be ignored as a parent directory is ignored.
     case: Case,
@@ -183,12 +169,8 @@ impl Ignore {
                     let ignore_blob = find(&attribute_files_in_index[idx].1, buf)
                         .map_err(|err| std::io::Error::new(std::io::ErrorKind::Other, err))?;
                     let ignore_path = gix_path::from_bstring(ignore_path_relative.into_owned());
-                    gix_glob::search::add_patterns_buffer(
-                        &mut self.stack.patterns,
-                        ignore_blob.data,
-                        ignore_path,
-                        Some(root),
-                    );
+                    self.stack
+                        .add_patterns_buffer(ignore_blob.data, ignore_path, Some(root));
                 }
                 Err(_) => {
                     // Need one stack level per component so push and pop matches.
@@ -197,120 +179,5 @@ impl Ignore {
             }
         }
         Ok(())
-    }
-}
-
-impl Attributes {
-    /// Create a new instance from an attribute match group that represents `globals`.
-    ///
-    /// A stack of attributes will be applied on top of it later.
-    pub fn new(globals: AttributeMatchGroup) -> Self {
-        Attributes {
-            globals,
-            stack: Default::default(),
-        }
-    }
-}
-
-impl From<AttributeMatchGroup> for Attributes {
-    fn from(group: AttributeMatchGroup) -> Self {
-        Attributes::new(group)
-    }
-}
-
-impl State {
-    /// Configure a state to be suitable for checking out files.
-    pub fn for_checkout(unlink_on_collision: bool, attributes: Attributes) -> Self {
-        State::CreateDirectoryAndAttributesStack {
-            unlink_on_collision,
-            #[cfg(debug_assertions)]
-            test_mkdir_calls: 0,
-            attributes,
-        }
-    }
-
-    /// Configure a state for adding files.
-    pub fn for_add(attributes: Attributes, ignore: Ignore) -> Self {
-        State::AttributesAndIgnoreStack { attributes, ignore }
-    }
-
-    /// Configure a state for status retrieval.
-    pub fn for_status(ignore: Ignore) -> Self {
-        State::IgnoreStack(ignore)
-    }
-}
-
-impl State {
-    /// Returns a vec of tuples of relative index paths along with the best usable OID for either ignore, attribute files or both.
-    ///
-    /// - ignores entries which aren't blobs
-    /// - ignores ignore entries which are not skip-worktree
-    /// - within merges, picks 'our' stage both for ignore and attribute files.
-    pub fn build_attribute_list(
-        &self,
-        index: &gix_index::State,
-        paths: &gix_index::PathStorageRef,
-        case: Case,
-    ) -> Vec<PathOidMapping> {
-        let a1_backing;
-        let a2_backing;
-        let names = match self {
-            State::IgnoreStack(v) => {
-                a1_backing = [(v.exclude_file_name_for_directories.as_bytes().as_bstr(), true)];
-                a1_backing.as_ref()
-            }
-            State::AttributesAndIgnoreStack { ignore, .. } => {
-                a2_backing = [
-                    (ignore.exclude_file_name_for_directories.as_bytes().as_bstr(), true),
-                    (".gitattributes".into(), false),
-                ];
-                a2_backing.as_ref()
-            }
-            State::CreateDirectoryAndAttributesStack { .. } => {
-                a1_backing = [(".gitattributes".into(), true)];
-                a1_backing.as_ref()
-            }
-        };
-
-        index
-            .entries()
-            .iter()
-            .filter_map(move |entry| {
-                let path = entry.path_in(paths);
-
-                // Stage 0 means there is no merge going on, stage 2 means it's 'our' side of the merge, but then
-                // there won't be a stage 0.
-                if entry.mode == gix_index::entry::Mode::FILE && (entry.stage() == 0 || entry.stage() == 2) {
-                    let basename = path
-                        .rfind_byte(b'/')
-                        .map(|pos| path[pos + 1..].as_bstr())
-                        .unwrap_or(path);
-                    let is_ignore = names.iter().find_map(|t| {
-                        match case {
-                            Case::Sensitive => basename == t.0,
-                            Case::Fold => basename.eq_ignore_ascii_case(t.0),
-                        }
-                        .then_some(t.1)
-                    })?;
-                    // See https://github.com/git/git/blob/master/dir.c#L912:L912
-                    if is_ignore && !entry.flags.contains(gix_index::entry::Flags::SKIP_WORKTREE) {
-                        return None;
-                    }
-                    Some((path.to_owned(), entry.id))
-                } else {
-                    None
-                }
-            })
-            .collect()
-    }
-
-    pub(crate) fn ignore_or_panic(&self) -> &Ignore {
-        match self {
-            State::IgnoreStack(v) => v,
-            State::AttributesAndIgnoreStack { ignore, .. } => ignore,
-            State::CreateDirectoryAndAttributesStack { .. } => {
-                unreachable!("BUG: must not try to check excludes without it being setup")
-            }
-        }
     }
 }
