+++ gix-worktree/tests/worktree/fs/cache/ignore.rs
@@ -106,7 +106,7 @@ fn check_against_baseline() -> crate::Result {
         ),
     );
     let paths_storage = index.take_path_backing();
-    let attribute_files_in_index = state.build_attribute_list(&index, &paths_storage, case);
+    let attribute_files_in_index = state.attribute_list_from_index(&index, &paths_storage, case);
     assert_eq!(
         attribute_files_in_index,
         vec![(
